Fixed the variable name and created exceptions file to capture exceptions
