+++ src/databricks/labs/remorph/reconcile/utils.py
@@ -1,2 +1,1 @@
-def alias_column_str(alias: str, columns: list[str]) -> list[str]:
-    return [f"{alias}.{column}" for column in columns]
+
